Move: split coin_select module
